Schema v8 cleanup: keep HighResBucket.T in seconds, retire INT_MS_FOLLOWUPS
Two small follow-ups on top of the v8 full-migration commit:

1. Revert HighResBucket.T from int32 ms to float64 seconds.

   The first sweep flipped HighResBucket.T to int32 ms for uniformity
   with the rest of the schema, but missed that this type is the v6
   legacy wire shape the frontend's bucket-fed panels read directly
   via getDefaultBuckets. mvd-web/static/app.js compares bucket.t
   against seconds-valued cursor times in ~20 places (timeline
   scrubber, key-moments hops, region-control spans, map playback,
   etc.); a unit flip there would silently break all of them.

   The clean fix: keep HighResBucket.T as float64 seconds — it's the
   public 'view-layer adapter' shape the frontend already reads, not a
   schema-stored field, and the design principle for v8 was 'public
   view-layer outputs stay seconds'. ToLegacyHighResBuckets emits
   vb.T directly with no conversion. Comment on the type spells out
   why this one timestamp stays in seconds.

   Verified by running view.Buckets and view.Events over a real demo
   on both dev and the current branch: events.T values match to ms
   precision (current emits clean -9.986 where dev had drift like
   -9.985999999999953); bucket output differs in 1 entry out of
   24,001, at the exact 528.15s window boundary — the same precision
   class the gib-respawn fix addressed.

2. Fold INT_MS_FOLLOWUPS.md into RESULT_SCHEMA.md and delete the
   standalone doc.

   The doc was created to capture deferred work and pros/cons of a
   future migration. With v8 finishing the migration in one cut, the
   deferred-work content no longer applies. The useful pieces — why
   integer ms, and the convention for adding new timestamped fields
   — moved into mvd-analytics/RESULT_SCHEMA.md as subsections of the
   existing v8 description, where future readers will find them.
   FOLLOWUPS.md cross-reference updated.

+++ mvd-analytics/result/timeline.go
@@ -78,10 +78,18 @@ type RegionStats struct {
 // HighResBucket is a compact bucket for high-resolution timeline data.
 // Uses short JSON keys to reduce payload size. Each bucket contains
 // per-player state snapshots and pre-computed team aggregations.
-// (Vestigial — at schema v7 buckets are produced on demand by
-// view.Buckets; this type survives only as a legacy adapter shape.)
+//
+// Vestigial — at schema v7 buckets are produced on demand by
+// view.Buckets; this type survives only as the wire shape the v6
+// frontend panels in mvd-web/static/app.js still consume via the
+// WASM bridge's getDefaultBuckets / ToLegacyHighResBuckets adapter.
+// T stays float64 seconds (the v6 shape) even though the rest of
+// schema v8 is int32 ms — the frontend compares bucket.t against
+// seconds-valued cursor times in ~20 panels, and a unit flip here
+// would force a parallel sweep of the legacy panel code. The
+// adapter does the ms→seconds conversion once at the boundary.
 type HighResBucket struct {
-	T  int32                         `json:"t"`            // Start time (ms, schema v8)
+	T  float64                       `json:"t"`            // Start time (float64 seconds, legacy frontend wire shape)
 	P  map[string]*HighResPlayerData `json:"p,omitempty"`  // Player data by name
 	TD map[string]*HighResTeamData   `json:"td,omitempty"` // Pre-computed team aggregations by team name
 }
